code structure improvements
