+++ frontend/src/components/Canvas/Canvas.js
@@ -1,18 +1,11 @@
-import React, { useRef, useEffect } from "react";
+import React from "react";
 import "./Canvas.scss";
 
 
-const Canvas = ({diagramRef, onDrop}) => {
-    function handleDrop(event) {
-        event.preventDefault();
-        const targetType = event.dataTransfer.getData('targetType');
-        console.log(targetType)
-        onDrop(targetType);
-      }
+const Canvas = ({ diagramRef }) => {
 
     return(
-        <div className="canvas-container" onDragOver={(e) => e.preventDefault()} onDrop={handleDrop}>
-            Canvas
+        <div className="canvas-container" onDragOver={(e) => e.preventDefault()} >
             <div ref={diagramRef} style={{ height: "100%", width: "100%"}}></div>
         </div>
     );
